100000_libdnf_utils_smartcols_Cell.hpp_use-private-scols_cell_get_alignment
Stretch's version of libsmartcols includes the (binary) symbol, but does not
export it via its API, probably to make the intermediate release
API-compatible.

libdnf needs it, so we'll just declare the function and hope that linking will
be able to pick it up just fine.

For newer versions, that should just turn into a comment.